v0.6.13: GPU Polyakov loop, NVK guard, cross-spring rewiring + docs cleanup
Core changes (barracuda v0.6.13):
- GPU-resident Polyakov loop shader (72x less data transfer than CPU readback)
- NVK allocation guard prevents nouveau PTE faults on Titan V
- su3_math_f64.wgsl: naga-safe SU(3) math for shader composition
- PRNG type-safety fix in su3_random_momenta_f64.wgsl
- 13/13 validation checks pass

New files:
- polyakov_loop_f64.wgsl, su3_math_f64.wgsl, su3_kinetic_energy_df64.wgsl,
  wilson_plaquette_df64.wgsl (DF64 shaders from toadStool S60)
- 7 new binaries: production_mixed_pipeline, validate_{three_substrate,
  adaptive_beta_scan, titan_oracle, multi_observable_npu, cg_predictor,
  online_learning}
- experiments/015 (mixed pipeline), 016 (cross-spring evolution map)
- whitePaper/baseCamp/ (per-domain research briefings)
- 3 wateringHole handoffs (v0.6.12, v0.6.13, toadStool absorption)

Docs cleanup:
- README, CONTROL_EXPERIMENT_STATUS, whitePaper/README updated to v0.6.13
- ABSORPTION_MANIFEST, EVOLUTION_READINESS brought current
- Resolved Polyakov loop TODO in experiment 015
- Cleared all stale version references (v0.6.11 → v0.6.13)
- Zero actionable TODO/FIXME remaining across .rs, .wgsl, and .md files

Made-with: Cursor

+++ barracuda/src/lattice/shaders/su3_random_momenta_f64.wgsl
@@ -39,8 +39,8 @@ fn box_muller_cos(u1: f64, u2: f64) -> f64 {
     var safe = u1;
     if safe < f64(1e-20) { safe = f64(1e-20); }
     let r = sqrt_f64(f64(-2.0) * log_f64(safe));
-    let theta = f32(f64(6.283185307179586) * u2);
-    return r * f64(cos(theta));
+    let theta = f64(6.283185307179586) * u2;
+    return r * cos(theta);
 }
 
 fn gaussian(link_idx: u32, pair: u32) -> f64 {
